+++ src/compiler/sema_decls.c
@@ -30,6 +30,7 @@ static bool sema_analyse_struct_members(SemaContext *context, Decl *decl);
 static inline bool sema_analyse_struct_member(SemaContext *context, Decl *parent, Decl *decl, bool *erase_decl);
 static inline bool sema_check_struct_holes(SemaContext *context, Decl *decl, Decl *member);
 static inline bool sema_analyse_bitstruct_member(SemaContext *context, Decl *parent, Decl *member, unsigned index, bool allow_overlap, bool *erase_decl);
+static bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *check_defined);
 
 static inline bool sema_analyse_doc_header(SemaContext *context, DeclId doc, Decl **params, Decl **extra_params, bool *pure_ref, bool is_raw_vaarg);
 
@@ -4827,6 +4828,7 @@ bool sema_analyse_var_decl_ct(SemaContext *context, Decl *decl, bool *check_defi
 			UNREACHABLE
 	}
 	if (check_defined) return true;
+	decl->resolve_status = RESOLVE_DONE;
 	return sema_add_local(context, decl);
 FAIL_CHECK:
 	if (check_defined)
@@ -4841,7 +4843,7 @@ FAIL:
 /**
  * Analyse a regular global or local declaration, e.g. int x = 123
  */
-bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *check_defined)
+static bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *check_defined)
 {
 	ASSERT(decl->decl_kind == DECL_VAR && "Unexpected declaration type");
 
@@ -4869,13 +4871,12 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 	// this should always be true.
 	if (!type_info && decl->is_extern)
 	{
-		SEMA_ERROR(decl, "A type is needed for the extern %s '%s'.", decl_to_name(decl), decl->name);
-		return decl_poison(decl);
+		RETURN_SEMA_ERROR(decl, "A type is needed for the extern %s '%s'.", decl_to_name(decl), decl->name);
 	}
 	ASSERT(type_info || decl->var.init_expr);
 
 	bool erase_decl = false;
-	if (!sema_analyse_attributes_for_var(context, decl, &erase_decl)) return decl_poison(decl);
+	if (!sema_analyse_attributes_for_var(context, decl, &erase_decl)) return false;
 
 	bool is_static = decl->var.is_static;
 	bool global_level_var = is_static || decl->var.kind == VARDECL_CONST || is_global;
@@ -4883,14 +4884,12 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 	if (decl->is_extern && decl->var.init_expr)
 	{
 		ASSERT(is_global);
-		SEMA_ERROR(decl->var.init_expr, "Extern globals may not have initializers.");
-		return decl_poison(decl);
+		RETURN_SEMA_ERROR(decl->var.init_expr, "Extern globals may not have initializers.");
 	}
 
 	if (decl->var.no_init && decl->var.init_expr)
 	{
-		SEMA_ERROR(decl->var.init_expr, "'@noinit' variables may not have initializers.");
-		return decl_poison(decl);
+		RETURN_SEMA_ERROR(decl->var.init_expr, "'@noinit' variables may not have initializers.");
 	}
 	if (erase_decl)
 	{
@@ -4909,14 +4908,12 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 		{
 			if (context->current_macro)
 			{
-				SEMA_ERROR(decl, "Macros with declarations may not be used outside of functions.");
-				return decl_poison(decl);
+				RETURN_SEMA_ERROR(decl, "Macros with declarations may not be used outside of functions.");
 			}
-			SEMA_ERROR(decl, "Variable declarations may not be used outside of functions.");
-			return decl_poison(decl);
+			RETURN_SEMA_ERROR(decl, "Variable declarations may not be used outside of functions.");
 		}
 		// Add a local to the current context, will throw error on shadowing.
-		if (!sema_add_local(context, decl)) return decl_poison(decl);
+		if (!sema_add_local(context, decl)) return false;
 	}
 
 
@@ -4928,77 +4925,69 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 		if (!init_expr)
 		{
 			ASSERT(kind == VARDECL_CONST);
-			SEMA_ERROR(decl, "Constants need to have an initial value.");
-			return decl_poison(decl);
+			RETURN_SEMA_ERROR(decl, "Constants need to have an initial value.");
 		}
 		ASSERT(!decl->var.no_init);
 		if (!check_defined && kind == VARDECL_LOCAL && !context_is_macro(context) && init_expr->expr_kind != EXPR_LAMBDA && !decl->var.safe_infer)
 		{
-			SEMA_ERROR(decl, "Defining a variable using 'var %s = ...' is only allowed inside a macro, or when defining a lambda. You can override this by adding the attribute '@safeinfer' to the declaration.", decl->name);
-			return decl_poison(decl);
+			RETURN_SEMA_ERROR(decl, "Defining a variable using 'var %s = ...' is only allowed inside a macro, or when defining a lambda. You can override this by adding the attribute '@safeinfer' to the declaration.", decl->name);
 		}
-		if (!sema_analyse_expr_rvalue(context, init_expr)) return decl_poison(decl);
+		if (!sema_analyse_expr_rvalue(context, init_expr)) return false;
 		if (check_defined || global_level_var || !type_is_aggregate(init_expr->type)) sema_cast_const(init_expr);
 		if (global_level_var && !expr_is_runtime_const(init_expr))
 		{
 			if (check_defined) return *check_defined = true, false;
-			SEMA_ERROR(init_expr, "This expression cannot be evaluated at compile time.");
-			return decl_poison(decl);
+			RETURN_SEMA_ERROR(init_expr, "This expression cannot be evaluated at compile time.");
 		}
 		decl->type = init_expr->type;
 		switch (sema_resolve_storage_type(context, init_expr->type))
 		{
 			case STORAGE_ERROR:
-				return decl_poison(decl);
+				return false;
 			case STORAGE_NORMAL:
 				break;
 			case STORAGE_WILDCARD:
 				if (check_defined) return *check_defined = true, false;
 				SEMA_ERROR(init_expr, "No type can be inferred from the optional result.");
-				return decl_poison(decl);
+				return false;
 			case STORAGE_VOID:
 				if (check_defined) return *check_defined = true, false;
-				SEMA_ERROR(init_expr, "You cannot initialize a value to 'void'.");
-				return decl_poison(decl);
+				RETURN_SEMA_ERROR(init_expr, "You cannot initialize a value to 'void'.");
 			case STORAGE_COMPILE_TIME:
 				if (check_defined) return *check_defined = true, false;
 				if (init_expr->type == type_untypedlist)
 				{
 					if (check_defined) return *check_defined = true, false;
-					SEMA_ERROR(init_expr,
+					RETURN_SEMA_ERROR(init_expr,
 					           "The type of an untyped list cannot be inferred, you can try adding an explicit type to solve this.");
-					return decl_poison(decl);
 				}
 				if (decl->var.kind == VARDECL_CONST)
 				{
-					SEMA_ERROR(init_expr,
+					RETURN_SEMA_ERROR(init_expr,
 					           "You cannot initialize a constant to %s, but you can assign the expression to a compile time variable.",
 					           type_invalid_storage_type_name(init_expr->type));
-					return decl_poison(decl);
 				}
-				SEMA_ERROR(init_expr, "You can't store a compile time type in a variable.");
-				return decl_poison(decl);
+				RETURN_SEMA_ERROR(init_expr, "You can't store a compile time type in a variable.");
 			case STORAGE_UNKNOWN:
 				if (check_defined) return *check_defined = true, false;
-				SEMA_ERROR(init_expr, "You cannot initialize a value to %s as it has unknown size.",
+				RETURN_SEMA_ERROR(init_expr, "You cannot initialize a value to %s as it has unknown size.",
 				           type_quoted_error_string(init_expr->type));
-				return decl_poison(decl);
 		}
 		if (!decl->alignment)
 		{
 			if (!sema_set_alloca_alignment(context, decl->type, &decl->alignment)) return false;
 		}
-		if (!sema_analyse_variable_type(context, decl->type, init_expr->loc)) return decl_poison(decl);
+		if (!sema_analyse_variable_type(context, decl->type, init_expr->loc)) return false;
 		// Skip further evaluation.
 		goto EXIT_OK;
 	}
 
 	if (!sema_resolve_type_info(context, type_info,
 	                            decl->var.init_expr ? RESOLVE_TYPE_ALLOW_INFER
-	                                                : RESOLVE_TYPE_DEFAULT)) return decl_poison(decl);
+	                                                : RESOLVE_TYPE_DEFAULT)) return false;
 
 	Type *type = decl->type = type_info->type;
-	if (!sema_analyse_variable_type(context, type, type_info->loc)) return decl_poison(decl);
+	if (!sema_analyse_variable_type(context, type, type_info->loc)) return false;
 
 	type = type_no_optional(type);
 	if (type_is_user_defined(type) && type->decl)
@@ -5009,16 +4998,15 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 
 	if (is_static && context->call_env.pure)
 	{
-		SEMA_ERROR(decl, "'@pure' functions may not have static variables.");
-		return decl_poison(decl);
+		RETURN_SEMA_ERROR(decl, "'@pure' functions may not have static variables.");
 	}
 
 	bool infer_len = type_len_is_inferred(decl->type);
 	if (!decl->var.init_expr && infer_len)
 	{
-		SEMA_ERROR(type_info, "The length cannot be inferred without an initializer.");
-		return decl_poison(decl);
+		RETURN_SEMA_ERROR(type_info, "The length cannot be inferred without an initializer.");
 	}
+	if (!infer_len) decl->resolve_status = RESOLVE_DONE;
 	if (decl->var.init_expr)
 	{
 		Expr *init = decl->var.init_expr;
@@ -5026,7 +5014,6 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 		if (!infer_len)
 		{
 			// Pre resolve to avoid problem with recursive definitions.
-			decl->resolve_status = RESOLVE_DONE;
 			if (!decl->alignment)
 			{
 				if (!sema_set_alloca_alignment(context, decl->type, &decl->alignment)) return false;
@@ -5062,8 +5049,7 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 			if (!expr_is_runtime_const(init))
 			{
 				if (check_defined) return *check_defined = true, false;
-				SEMA_ERROR(init, "The expression must be a constant value.");
-				return decl_poison(decl);
+				RETURN_SEMA_ERROR(init, "The expression must be a constant value.");
 			}
 		}
 		if (!success) goto EXIT_OK;
@@ -5074,6 +5060,7 @@ bool sema_analyse_var_decl(SemaContext *context, Decl *decl, bool local, bool *c
 		}
 	}
 	EXIT_OK:;
+	decl->resolve_status = RESOLVE_DONE;
 	// Patch the external name for local consts and static variables.
 	if ((decl->var.kind == VARDECL_CONST || is_static) && !decl->extname && context->call_env.kind == CALL_ENV_FUNCTION)
 	{
@@ -5730,6 +5717,27 @@ TYPE_MISMATCH:
 		type_quoted_error_string(replacement->type), type_quoted_error_string(replaced->type));
 	SEMA_NOTE(replaced, "The replaced definition was here.");
 	return false;
+}
+
+bool sema_analyse_local(SemaContext *context, Decl *decl, bool *failed_ref)
+{
+	if (decl->resolve_status == RESOLVE_DONE) return decl_ok(decl);
+	if (decl->resolve_status == RESOLVE_RUNNING)
+	{
+		SEMA_ERROR(decl, decl->name
+			? "Recursive definition of '%s'."
+			: "Recursive definition of anonymous declaration.", decl->name);
+		return decl_poison(decl);
+	}
+	decl->resolve_status = RESOLVE_RUNNING;
+	assert(decl->decl_kind == DECL_VAR);
+	if (!sema_analyse_var_decl(context, decl, true, failed_ref))
+	{
+		if (failed_ref && *failed_ref) return false;
+		return decl_poison(decl);
+	}
+	ASSERT(failed_ref || decl->resolve_status == RESOLVE_DONE);
+	return true;
 }
 
 bool sema_analyse_decl(SemaContext *context, Decl *decl)